refactor: 아키텍처 구조 개선 - domains, shared 폴더 적용
