--- src/basic/components/index.js
@@ -1,18 +1,0 @@
-// ==================== 순수 HTML 구조 컴포넌트들 ====================
-
-// Layout 컴포넌트들
-export { Header } from './Header.js';
-export { MainGrid, LeftColumn } from './Layout.js';
-export { OrderSummary } from './OrderSummary.js';
-export { HelpToggleButton, HelpManualOverlay } from './HelpManual.js';
-
-// 상품 관련 컴포넌트들
-export { ProductPanel } from './ProductPanel.js';
-export { ProductSelect } from './ProductSelect.js';
-export { AddToCartButton } from './AddToCartButton.js';
-export { StockInfo } from './StockInfo.js';
-export { ProductOption } from './ProductSelector.js';
-
-// 장바구니 관련 컴포넌트들
-export { CartContainer } from './CartContainer.js';
-export { CartItem } from './CartItem.js';
