+++ src/basic/domains/index.js
@@ -1,0 +1,19 @@
+// Cart 도메인
+export * as CartController from './cart/CartController.js';
+export * as CartService from './cart/CartService.js';
+export * as CartRenderer from './cart/CartRenderer.js';
+export * as CartComponents from './cart/components/index.js';
+
+// Product 도메인
+export * as ProductController from './product/ProductController.js';
+export * as StockService from './product/StockService.js';
+export * as ProductRenderer from './product/ProductRenderer.js';
+export * as ProductComponents from './product/components/index.js';
+
+// Order 도메인
+export * as DiscountService from './order/DiscountService.js';
+export * as PointService from './order/PointService.js';
+
+// Promotion 도메인
+export * as LightningSaleService from './promotion/LightningSaleService.js';
+export * as SuggestedSaleService from './promotion/SuggestedSaleService.js';
