+++ src/basic/domains/order/PointService.js
@@ -5,9 +5,9 @@ import {
   KEYBOARD_ID,
   MOUSE_ID,
   MONITOR_ID,
-} from '../constants/index.js';
+} from '../../shared/constants/index.js';
 
-// ==================== 포인트 계산 서비스 ====================
+// ==================== 주문 포인트 계산 서비스 ====================
 
 /**
  * 기본 포인트 계산 (1,000원당 1포인트)
@@ -38,18 +38,18 @@ export function applyTuesdayBonus(basePoints, isTuesday) {
 
 /**
  * 장바구니 상품 타입 확인
- * @param {Array} cartItems - 장바구니 아이템들
+ * @param {Array} cartData - 장바구니 데이터 [{id, quantity}]
  * @param {Function} getProductById - 상품 조회 함수
  * @returns {Object} { hasKeyboard, hasMouse, hasMonitor }
  */
-export function checkProductTypes(cartItems, getProductById) {
+export function checkProductTypes(cartData, getProductById) {
   const types = {
     hasKeyboard: false,
     hasMouse: false,
     hasMonitor: false,
   };
 
-  for (const cartItem of cartItems) {
+  for (const cartItem of cartData) {
     const product = getProductById(cartItem.id);
     if (!product) continue;
 
@@ -124,14 +124,14 @@ export function calculateQuantityBonus(totalQuantity) {
  * 전체 포인트 계산 (메인 오케스트레이터)
  * @param {number} finalAmount - 최종 결제 금액
  * @param {number} totalQuantity - 총 수량
- * @param {Array} cartItems - 장바구니 아이템들
+ * @param {Array} cartData - 장바구니 데이터 [{id, quantity}]
  * @param {Function} getProductById - 상품 조회 함수
  * @returns {Object} { finalPoints, pointsDetails }
  */
 export function calculateTotalPoints(
   finalAmount,
   totalQuantity,
-  cartItems,
+  cartData,
   getProductById
 ) {
   // 1. 기본 포인트 계산
@@ -151,7 +151,7 @@ export function calculateTotalPoints(
   pointsDetails.push(...tuesdayResult.details);
 
   // 3. 세트 구매 보너스
-  const productTypes = checkProductTypes(cartItems, getProductById);
+  const productTypes = checkProductTypes(cartData, getProductById);
   const setBonusResult = calculateSetBonus(productTypes);
   finalPoints += setBonusResult.bonusPoints;
   pointsDetails.push(...setBonusResult.bonusDetails);
