+++ src/basic/domains/product/components/ProductPanel.js
