+++ src/basic/domains/product/components/StockInfo.js
