+++ src/basic/domains/product/components/index.js
@@ -1,0 +1,5 @@
+export { ProductPanel } from './ProductPanel.js';
+export { ProductSelect } from './ProductSelect.js';
+export { ProductOption } from './ProductSelector.js';
+export { AddToCartButton } from './AddToCartButton.js';
+export { StockInfo } from './StockInfo.js';
