+++ src/basic/shared/components/index.js
@@ -1,0 +1,3 @@
+export { Header } from './Header.js';
+export { MainGrid, LeftColumn } from './Layout.js';
+export { HelpToggleButton, HelpManualOverlay } from './HelpManual.js';
