+++ src/basic/shared/core/index.js
@@ -1,5 +1,3 @@
-// ==================== 코어 레이어 통합 export ====================
-
 // 비즈니스 상태 관리
 export {
   getProducts,
