+++ src/basic/shared/utils/formatters.js
@@ -1,4 +1,3 @@
-// ==================== 가격 포맷팅 유틸리티 함수들 ====================
 export function formatPrice(price) {
   return '₩' + price.toLocaleString();
 }
